requirements.txt
requests
beautifulsoup4
python-docx
PyMuPDF
playwright